PHP 7.4.30 Swoole 4.8.12

+++ ext/mysqli/tests/mysqli_pam_sha256_public_key_option.phpt
@@ -1,9 +1,9 @@
 --TEST--
 PAM: SHA-256, option: MYSQLI_SERVER_PUBLIC_KEY
---EXTENSIONS--
-mysqli
 --SKIPIF--
 <?php
+require_once('skipif.inc');
+require_once('skipifemb.inc');
 require_once('skipifconnectfailure.inc');
 
 ob_start();
@@ -55,8 +55,10 @@ if (strlen($row['Value']) != fwrite($fp, $row['Value'])) {
     die(sprintf("skip Failed to create pub key file"));
 }
 
-// Ignore errors because this variable exists only in MySQL 5.6 and 5.7
-$link->query("SET @@session.old_passwords=2");
+
+if (!$link->query("SET @@session.old_passwords=2")) {
+    die(sprintf("skip Cannot set @@session.old_passwords=2 [%d] %s", $link->errno, $link->error));
+}
 
 $link->query('DROP USER shatest');
 $link->query("DROP USER shatest@localhost");
@@ -67,8 +69,8 @@ if (!$link->query('CREATE USER shatest@"%" IDENTIFIED WITH sha256_password') ||
     die(sprintf("skip CREATE USER failed [%d] %s", $link->errno, $link->error));
 }
 
-if (!$link->query('SET PASSWORD FOR shatest@"%" = "shatest"') ||
-    !$link->query('SET PASSWORD FOR shatest@"localhost" = "shatest"')) {
+if (!$link->query('SET PASSWORD FOR shatest@"%" = PASSWORD("shatest")') ||
+    !$link->query('SET PASSWORD FOR shatest@"localhost" = PASSWORD("shatest")')) {
     die(sprintf("skip SET PASSWORD failed [%d] %s", $link->errno, $link->error));
 }
 
@@ -84,7 +86,6 @@ if (!$link->query(sprintf("GRANT SELECT ON TABLE %s.test TO shatest@'%%'", $db))
 }
 
 $link->close();
-echo "nocache";
 ?>
 --FILE--
 <?php
